@@ -15,13 +15,12 @@
         "discard-changes": true
     },
     "require": {
-        "php": "^8.0"
+        "php": "^8.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.6",
-        "friendsofphp/php-cs-fixer": "^3.15",
-        "phan/phan": "^5.4",
-        "php-parallel-lint/php-parallel-lint": "^1.3"
+        "php-parallel-lint/php-parallel-lint": "^1.4",
+        "phpunit/phpunit": "^11",
+        "friendsofphp/php-cs-fixer": "^3.88"
     },
     "autoload": {
         "psr-4": {
@@ -40,13 +39,9 @@
         "unit": [
             "phpunit --colors=never"
         ],
-        "phan": [
-            "phan -p"
-        ],
         "test": [
             "@lint",
-            "@unit",
-            "@phan"
+            "@unit"
         ],
         "fix": [
             "php-cs-fixer fix --config .php-cs-fixer.dist.php src tests"
